+++ particular/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Built-in compute methods implemented with the different provided storage structs.
 - `ComputeMethod` no longer generic over `V`.
 - Renamed `WgpuData` to `WgpuResources` and `wgpu_data` to `gpu_compute`.
+- `gpu_compute` (internals of `gpu::BruteForce`) is public.
 - `gpu::BruteForce` holds a reference to a `wgpu::Device`, `wgpu::Queue` and a mutable reference to a `GpuData`.
 - Renamed `BruteForce` compute methods to `BruteForceScalar` for better differentiation with simd variant.
 - `BruteForceSIMD` generic over the lane count.